requirements.txt
python-dotenv
google-generativeai
pytest